# [SPARK-22010][PySpark] Change fromInternal method of TimestampType