--- a/python/pyspark/sql/types.py
+++ b/python/pyspark/sql/types.py
@@ -19,7 +19,7 @@ import sys
 import decimal
 import time
 import datetime
-import dateutil
+import dateutil.tz
 import calendar
 import json
 import re