--- a/python/pyspark/sql/types.py
+++ b/python/pyspark/sql/types.py
@@ -200,7 +200,8 @@ class TimestampType(AtomicType):
     def fromInternal(self, ts):
         if ts is not None:
             # using int to avoid precision loss in float
-            y, m, d, hh, mm, ss, _, _, _ = time.gmtime(ts // 1000000) if _is_utc else time.localtime(ts // 1000000)
+            y, m, d, hh, mm, ss, _, _, _ = (time.gmtime(ts // 1000000) if _is_utc
+                                            else time.localtime(ts // 1000000))
             return datetime.datetime(y, m, d, hh, mm, ss, ts % 1000000)