@maver1ck 2017-09-18
Any idea why we're not using
`datetime.datetime.fromtimestamp(ts / 100000.)` ?
There is a comment about overflow. But if it exists ?