@maver1ck 2017-09-18
I'm asking because such a code is 2x faster than my solution