@HyukjinKwon 2017-09-18
I think it was changed back in https://github.com/apache/spark/pull/7363. I think that avoids precision loss from Python's `float` limitation:

```python
>>> decimal.Decimal(1 / 1e6)
Decimal('9.99999999999999954748111825886258685613938723690807819366455078125E-7')
>>> decimal.Decimal(1 / 100000)
Decimal('0')
>>> decimal.Decimal(1 / 100000.)
Decimal('0.000010000000000000000818030539140313095458623138256371021270751953125')
```

Also, I think `//` is used for Python 3 behaviour:

```python
>>> 5 / 2
2.5
>>> 5 // 2
2
```