@HyukjinKwon 2017-09-18
Seems fine to me too as is. @maver1ck, I think you could take out `[WIP]` and let it be merged.