@maver1ck 2017-09-18
OK. It passed all tests, so let merge it