@maver1ck 2017-09-21
It was introduced with this PEP.
https://www.python.org/dev/peps/pep-0495/