@HyukjinKwon 2017-11-06
@maver1ck, if you'd like to leave it as is, I think we should better close it.