@maver1ck 2017-11-07
I think that porting changes from Python 3.6 give us too complicated code.
I'm closing it.